Refactor code to remove unused imports and functions
